+++ online-marketing-back/src/main/java/com/diplomski/onlinemarketing/service/StoreService.java
@@ -3,5 +3,5 @@ package com.diplomski.onlinemarketing.service;
 import com.diplomski.onlinemarketing.entity.Store;
 import com.diplomski.onlinemarketing.service.generic.GenericService;
 
-public interface StoreService extends GenericService<Store, Long> {
+public interface StoreService extends GenericService<Store, Integer> {
 }
