+++ online-marketing-back/src/main/java/com/diplomski/onlinemarketing/service/UserService.java
@@ -3,5 +3,5 @@ package com.diplomski.onlinemarketing.service;
 import com.diplomski.onlinemarketing.entity.User;
 import com.diplomski.onlinemarketing.service.generic.GenericService;
 
-public interface UserService extends GenericService<User, Long> {
+public interface UserService extends GenericService<User, Integer> {
 }
